+++ include/parlay/internal/scheduler_plugins/tbb.h
@@ -16,7 +16,7 @@ inline size_t num_workers() { return tbb::this_task_arena::max_concurrency(); }
 inline size_t worker_id() { return tbb::task_arena::current_thread_index(); }
 
 template <class F>
-inline void parallel_for(size_t start, size_t end, F f, size_t granularity, bool) {
+inline void parallel_for(size_t start, size_t end, F f, long granularity, bool) {
   if (granularity == 0) {
     tbb::parallel_for(tbb::blocked_range<size_t>(start, end), [&](const tbb::blocked_range<size_t>& r) {
       for (auto i = r.begin(); i != r.end(); ++i) {
